pyproject.toml
[project]
name = "fyp-project"
version = "0.1.0"
description = "Hybrid Sentiment and Technical Indicator-Based Trading Strategy Using Machine Learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backtesting>=0.6.5",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.12",
    "finnhub-python>=2.4.25",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "pandas-ta>=0.3.14b1",
    "plotly>=6.0.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.0",
    "shap>=0.46.0",
    "ta-lib>=0.6.8",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "xgboost>=2.1.0",
    "yfinance>=0.2.66",
]

[project.optional-dependencies]
extras = [
    "gliner>=0.2.0",
    "sentence-transformers>=3.0.0",
]
